+++ src/components/App.css
@@ -1,0 +1,7 @@
+html {
+  background-color: black;
+}
+.App {
+  text-align: center;
+  color: white;
+}
